fix more reformating "bugs"
