+++ Complexity_and_Scope/TodoAgility.Agile/CQRS/CommandHandlers/ChangeActivityStatusCommandHandler.cs
@@ -18,7 +18,7 @@
 
 using TodoAgility.Agile.CQRS.CommandHandlers.Framework;
 using TodoAgility.Agile.Domain.Aggregations;
-using TodoAgility.Agile.Domain.DomainEvents.Framework;
+using TodoAgility.Agile.Domain.Framework.DomainEvents;
 using TodoAgility.Agile.Persistence.Framework;
 using TodoAgility.Agile.Persistence.Repositories;
