+++ Complexity_and_Scope/TodoAgility.Agile/Domain/BusinessObjects/Description.cs
@@ -33,9 +33,15 @@ namespace TodoAgility.Agile.Domain.BusinessObjects
 
         public bool Equals(Description other)
         {
-            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(null, other))
+            {
+                return false;
+            }
 
-            if (ReferenceEquals(this, other)) return true;
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
 
             return string.Equals(_description, other._description);
         }
@@ -48,14 +54,19 @@ namespace TodoAgility.Agile.Domain.BusinessObjects
         public static Description From(string description)
         {
             if (string.IsNullOrEmpty(description) || string.IsNullOrWhiteSpace(description))
+            {
                 throw new ArgumentException("A descrição informada é nulo, vazio ou composto por espaços em branco.",
                     nameof(description));
+            }
+
 
 
             if (description.Length > DESCRIPTION_LENGTH_LIMIT)
+            {
                 throw new ArgumentException(
                     $"A descripção excedeu o limite máximo de {DESCRIPTION_LENGTH_LIMIT} definido.",
                     nameof(description));
+            }
 
 
             return new Description(description);
@@ -63,11 +74,20 @@ namespace TodoAgility.Agile.Domain.BusinessObjects
 
         public override bool Equals(object obj)
         {
-            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(null, obj))
+            {
+                return false;
+            }
 
-            if (ReferenceEquals(this, obj)) return true;
+            if (ReferenceEquals(this, obj))
+            {
+                return true;
+            }
 
-            if (obj.GetType() != GetType()) return false;
+            if (obj.GetType() != GetType())
+            {
+                return false;
+            }
 
             return Equals((Description) obj);
         }
