+++ Complexity_and_Scope/TodoAgility.Agile/Domain/DomainEvents/ProjectAggregateActivityAddedHandler.cs
@@ -2,8 +2,8 @@ using System.Collections.Generic;
 using LiteDB;
 using Microsoft.EntityFrameworkCore;
 using TodoAgility.Agile.Domain.BusinessObjects;
-using TodoAgility.Agile.Domain.DomainEvents.Framework;
 using TodoAgility.Agile.Domain.Framework.BusinessObjects;
+using TodoAgility.Agile.Domain.Framework.DomainEvents;
 using TodoAgility.Agile.Persistence.Framework;
 using TodoAgility.Agile.Persistence.Model;
 using TodoAgility.Agile.Persistence.Repositories;
