+++ Complexity_and_Scope/TodoAgility.Agile/Domain/Framework/BusinessObjects/ValueObject.cs
@@ -48,16 +48,35 @@ namespace TodoAgility.Agile.Domain.Framework.BusinessObjects
 
         public bool Equals(TValueObject other)
         {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
+            if (ReferenceEquals(null, other))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+            
             return GetValueComponents().SequenceEqual(other.GetValueComponents());
         }
 
         public override bool Equals(object obj)
         {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != GetType()) return false;
+            if (ReferenceEquals(null, obj))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, obj))
+            {
+                return true;
+            }
+
+            if (obj.GetType() != GetType())
+            {
+                return false;
+            }
 
             var vo = (ValueObject<TValueObject>) obj;
             return Equals(vo);
